src/backend/requirements.txt
flask
pdf2image
pillow
pytest
pytest-cov
numpy
opencv-contrib-python
tensorflow
pandas
dataclasses-json
